Cleanup library loading

+++ src/main/java/jni/LibraryLoader.java
@@ -1,4 +1,4 @@
-// from: https://github.com/stwind/gradle-jni
+// Based on: https://github.com/stwind/gradle-jni
 
 package jni;
 
@@ -6,41 +6,45 @@ import java.io.*;
 import java.net.URL;
 
 public class LibraryLoader {
-    public static String getExt() {
+    /**
+     * @return The appropriate dynamic library file extension for the operating system
+     */
+    private static String getExt() {
         String osName = System.getProperty("os.name");
-        if (osName.equals("Linux"))
-            return "so";
-        else if (osName.equals("Mac OS X"))
-            return "dylib";
-        else
+        if (osName.startsWith("Windows")) {
             return "dll";
+        }
+        return switch (osName) {
+            case "Linux" -> "so";
+            case "Mac OS X" -> "dylib";
+            default -> throw new UnsupportedOperationException("Unknown operating system");
+        };
     }
 
-    public static Boolean load(Class<?> cls, String name) {
+    /**
+     * Loads a JNI dynamic library
+     * @param cls Class to load native methods
+     * @param name Name of the library
+     */
+    public static void load(Class<?> cls, String name) {
         String path = "/lib" + name + "." + getExt();
         URL url = cls.getResource(path);
-
-        Boolean success = false;
+        if (url == null) {
+            throw new RuntimeException("Library " + name + " not found");
+        }
         try {
             final File libfile = File.createTempFile(name, ".lib");
             libfile.deleteOnExit();
 
-            final InputStream in = url.openStream();
-            final OutputStream out = new BufferedOutputStream(new FileOutputStream(libfile));
-
-            int len = 0;
-            byte[] buffer = new byte[8192];
-            while ((len = in.read(buffer)) > -1)
-                out.write(buffer, 0, len);
-            out.close();
-            in.close();
+            try (InputStream in = url.openStream()) {
+                try (FileOutputStream out = new FileOutputStream(libfile)) {
+                    in.transferTo(out);
+                }
+            }
 
             System.load(libfile.getAbsolutePath());
-            success = true;
-        } catch (IOException x) {
-
+        } catch (IOException e) {
+            throw new RuntimeException(e);
         }
-
-        return success;
     }
 }
